add Rating typealias

+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/rating/database/PostsLikesMessagesTable.kt
@@ -110,7 +110,7 @@ class PostsLikesMessagesTable(
             selectAll().mapNotNull {
                 it[postId]
             }.map {
-                it.toLong() to postsLikesTable.getPostRating(it)
+                it.toLong() to postsLikesTable.getPostRating(it).toFloat()
             }
         }
     }
